% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_interneuron}
\alias{classify_interneuron}
\title{Classify an interneruon}
\usage{
classify_interneuron(file, layer)
}
\arguments{
\item{file}{a character. Path to the reconstruction SWC file.}

\item{layer}{a character. The layer containing the soma.}
}
\description{
Classify an interneruon
}
